Add JSON parser on client side

+++ Server/Game/Vector.cs
@@ -6,11 +6,11 @@ namespace Server.Game
     public class Vector
     {
         [DataMember]
-        public int X { get; set; }
+        public double X { get; set; }
         [DataMember]
-        public int Y { get; set; }
+        public double Y { get; set; }
 
-        public Vector(int x, int y)
+        public Vector(double x, double y)
         {
             X = x;
             Y = y;
